websockets/constructor/002.html

<!doctype html>
<title>WebSockets: new WebSocket(invalid url)</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
var domain = SCHEME_DOMAIN_PORT.split('://')[1];

test(function() {assert_throws("SyntaxError", function(){new WebSocket("/test")})});
test(function() {assert_throws("SyntaxError", function(){new WebSocket("ws://foo bar.com/")})});
test(function() {assert_throws("SyntaxError", function(){new WebSocket("wss://foo bar.com/")})});
test(function() {assert_throws("SyntaxError", function(){new WebSocket("http://"+domain+"/")})});
test(function() {assert_throws("SyntaxError", function(){new WebSocket("mailto:example@example.org")})});
test(function() {assert_throws("SyntaxError", function(){new WebSocket("about:blank")})});
test(function() {assert_throws("SyntaxError", function(){new WebSocket("ws://"+domain+"/#")})});
test(function() {assert_throws("SyntaxError", function(){new WebSocket("ws://"+domain+"/#test")})});
test(function() {assert_throws("SyntaxError", function(){new WebSocket("?test")})});
test(function() {assert_throws("SyntaxError", function(){new WebSocket("#test")})});
</script>
