LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/constructor/002-expected.txt

CONSOLE MESSAGE: Wrong url scheme for WebSocket http://web-platform.test:8800/test
CONSOLE MESSAGE: Invalid url for WebSocket ws://foo bar.com/
CONSOLE MESSAGE: Invalid url for WebSocket wss://foo bar.com/
CONSOLE MESSAGE: Wrong url scheme for WebSocket http://web-platform.test:8800/
CONSOLE MESSAGE: Wrong url scheme for WebSocket mailto:example@example.org
CONSOLE MESSAGE: Wrong url scheme for WebSocket about:blank
CONSOLE MESSAGE: URL has fragment component ws://web-platform.test:49001/#
CONSOLE MESSAGE: URL has fragment component ws://web-platform.test:49001/#test
CONSOLE MESSAGE: Wrong url scheme for WebSocket http://web-platform.test:8800/websockets/constructor/002.html?test
CONSOLE MESSAGE: Wrong url scheme for WebSocket http://web-platform.test:8800/websockets/constructor/002.html#test

PASS WebSockets: new WebSocket(invalid url)
PASS WebSockets: new WebSocket(invalid url) 1
PASS WebSockets: new WebSocket(invalid url) 2
PASS WebSockets: new WebSocket(invalid url) 3
PASS WebSockets: new WebSocket(invalid url) 4
PASS WebSockets: new WebSocket(invalid url) 5
PASS WebSockets: new WebSocket(invalid url) 6
PASS WebSockets: new WebSocket(invalid url) 7
PASS WebSockets: new WebSocket(invalid url) 8
PASS WebSockets: new WebSocket(invalid url) 9
